include/server/requirements.txt
tornado
cython
futures
riak
schematics
scipy
pyzmq
Jinja2
arrow
pycurl
ujson
